experiment_control/428hub/photonmover/requirements.txt
asn1crypto==1.2.0
backcall==0.1.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
colorama==0.4.3
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
idna==2.8
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
kiwisolver==1.1.0
matplotlib==3.1.2
menuinst==1.4.16
nidaqmx==0.5.7
numpy==1.17.4
parso==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.2
pycosat==0.6.3
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.0.0
pyparsing==2.4.5
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.1
PyVISA==1.10.1
pywin32==223
requests==2.22.0
ruamel-yaml==0.15.46
scipy==1.3.3
six==1.12.0
tqdm==4.36.1
traitlets==4.3.3
urllib3==1.24.2
wcwidth==0.1.7
win-inet-pton==1.1.0
wincertstore==0.2
wxPython==4.0.7.post2
